Decouple onliners presence from LiveView socket
