+++ lib/chat_web/live/main_live/page/lobby.ex
@@ -49,7 +49,7 @@ defmodule ChatWeb.MainLive.Page.Lobby do
     |> Page.Dialog.store_room_key_copy(new_room_identity)
     |> Page.Login.store_new_room(new_room_identity)
     |> assign_room_list()
-    |> Page.OnlinersPresence.update()
+    |> Page.Shared.update_onliners_presence()
     |> Page.Room.init({new_room_identity, new_room})
   end
