+++ lib/chat_web/live/main_live/page/shared.ex
@@ -1,7 +1,10 @@
 defmodule ChatWeb.MainLive.Page.Shared do
   @moduledoc "Shared page functions"
 
+  alias Chat.Identity
+  alias ChatWeb.OnlinersPresence
   alias Phoenix.HTML.Safe
+  alias Phoenix.LiveView.Socket
 
   def mime_type(nil), do: "application/octet-stream"
   def mime_type(""), do: mime_type(nil)
@@ -20,4 +23,38 @@ defmodule ChatWeb.MainLive.Page.Shared do
     |> Safe.to_iodata()
     |> IO.iodata_to_binary()
   end
+
+  def track_onliners_presence(%Socket{} = socket) do
+    {:ok, _} =
+      OnlinersPresence.track(socket.root_pid, presence_key(socket), get_user_keys(socket))
+
+    socket
+  end
+
+  def update_onliners_presence(%Socket{} = socket) do
+    OnlinersPresence.update(socket.root_pid, presence_key(socket), get_user_keys(socket))
+
+    socket
+  end
+
+  def untrack_onliners_presence(%Socket{} = socket) do
+    OnlinersPresence.untrack(socket.root_pid, presence_key(socket))
+
+    socket
+  end
+
+  defp presence_key(%Socket{assigns: %{me: me}}) do
+    Enigma.hash(me)
+  end
+
+  defp get_user_keys(%Socket{assigns: %{me: me, room_map: room_map}})
+       when not is_nil(me) and not is_nil(room_map) do
+    room_map
+    |> Stream.map(fn {_key, room} -> room end)
+    |> Stream.concat([me])
+    |> Enum.map(&Identity.pub_key/1)
+    |> MapSet.new()
+  end
+
+  defp get_user_keys(_socket), do: MapSet.new()
 end
